+++ packages/extension/src/view/devtools/pages/privateAdvertising/attributionReporting/activeSources/dummyData.json
