+++ packages/extension/src/view/devtools/pages/privateAdvertising/attributionReporting/index.tsx
