+++ packages/extension/src/view/devtools/pages/privateAdvertising/attributionReporting/panel.tsx
