+++ packages/extension/src/view/devtools/pages/privateAdvertising/attributionReporting/sourceRegistrations/index.tsx
