+++ packages/extension/src/view/devtools/pages/privateAdvertising/attributionReporting/triggerRegistrations/index.tsx
