+++ packages/extension/src/view/devtools/pages/privateAdvertising/attributionReporting/utils/calculateRegistrationDate.ts
